merge: Fixed std items on create invoice page (#55)
* #SC-100 Fix error in delete items in invoice creation

* SC-64 Fix invoice page bugs

* SC-93 Fix std items on create invoice page

* SC-93 Fix std items on create invoice page

* SC-93 ​Items on standard list should be selected if exist in items table

* bug: Remove items from table after deselecting items in fees list.

* fix: SC-93 Enable toggle behavior in standard item list with elements in the table

---------

+++ lib/data/invoice/model/invoice_request.dart
@@ -184,7 +184,7 @@ class PaymentScheduleRequest {
 class InvoiceItemRequest {
   final int quantity;
   final String description;
-  final int? grossPrice;
+  final double? grossPrice;
   final int? netPrice;
   final int? taxAmount;
   final double? taxRate;
